Move conftest.py to root dir to try to avoid segmentation fault issue
